homematicip_cloud: address review feedback
- Move migration code to migration.py (keep __init__.py small)
- Remove _async_remove_obsolete_entities (dead code after migration)
- Make feature_id required (str, not str | None)
- Remove legacy unique_id fallback
- Set _feature_id as class attribute on alarm_control_panel
